app/requirements.txt
Django==3.2
djangorestframework==3.14.0
psycopg2-binary
gunicorn
drf-yasg
django-filter==21.1
djangorestframework-simplejwt
